requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
elasticsearch==7.17.9
python-dotenv==1.0.0
pydantic==1.10.8
python-multipart==0.0.6
loguru==0.7.0  
pytest==7.4.0  
pytest-asyncio==0.21.0  
pytest-cov==4.1.0  
httpx==0.24.1  